// src/payload-types.ts
export interface Media {
  id: number;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "todos".
 */
export interface Todo {
  id: number;
  title: string;
  description?: string | null;
  completed?: boolean | null;
  media?: (number | null) | Media;
  createdAt: string;
  updatedAt: string;
}